♻️ 関数分け&説明追加
