+++ src/data/parse_redirects.c
@@ -6,47 +6,11 @@
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/22 22:35:26 by miyuu             #+#    #+#             */
-/*   Updated: 2025/02/28 22:09:30 by miyuu            ###   ########.fr       */
+/*   Updated: 2025/02/28 23:03:51 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
-
-int	has_from_fd(char *cmds_text, int i)
-{
-	int	m;
-
-	m = i;
-	while (i > 0 && ft_isdigit(cmds_text[i - 1]))
-		i--;
-	if (i < m)
-		return (ft_atoi(&cmds_text[i]));
-	if (cmds_text[i] == '<')
-		return (0);
-	else if (cmds_text[i] == '>')
-		return (1);
-	return (-1);
-}
-
-char	*get_redirect_path(char *redir_symbol, char *next_word)
-{
-	char	*next;
-	size_t	len;
-	char	*path;
-
-	while (*redir_symbol && (*redir_symbol == '>' || *redir_symbol == '<'))
-		redir_symbol++;
-	if (!*redir_symbol)
-		return (next_word);
-
-	next = has_redirect(redir_symbol);
-	if (!next)
-		return (redir_symbol);
-
-	len = next - redir_symbol;
-	path = ft_strndup(redir_symbol, len);
-	return (path);
-}
 
 /*
  * Function:parse_redirects
@@ -69,7 +33,7 @@ void	parse_redirects(t_redirect **redir, t_heredoc **hd, \
 	redir_symbol = has_redirect(word);
 	while (redir_symbol)
 	{
-		from_fd = has_from_fd(word, redir_symbol - word);
+		from_fd = get_redirect_from_fd(word, redir_symbol - word);
 		path = get_redirect_path(redir_symbol, next_word);
 		if ((ft_strncmp(redir_symbol, "<<", 2)) == 0)
 		{
@@ -81,7 +45,6 @@ void	parse_redirects(t_redirect **redir, t_heredoc **hd, \
 			add_struct_redirect(redir, REDIR_IN, from_fd, path);
 		else if ((ft_strncmp(redir_symbol, ">", 1)) == 0)
 			add_struct_redirect(redir, REDIR_OUT, from_fd, path);
-
 		while (*redir_symbol && (*redir_symbol == '>' || *redir_symbol == '<'))
 			redir_symbol++;
 		redir_symbol = has_redirect(redir_symbol);
